Rearranged the pseudocode, added a little more description and initialized cabal in a way that makes it more readbale and easier to manage packages
