+++ dapp3602022-t4.cabal
@@ -14,6 +14,7 @@ version:            0.1.0.0
 -- The license under which the package is released.
 -- license:
 author:             Alexander Watanabe
+contributors:       Otobong Peter & Augusta Erhibolo
 maintainer:         alex.watanabe@gmail.com
 
 -- A copyright notice.
