+++ dist-newstyle/cache/plan.json
@@ -1,0 +1,1 @@
+{"cabal-version":"3.6.2.0","cabal-lib-version":"3.6.2.0","compiler-id":"ghc-8.10.7","os":"windows","arch":"x86_64","install-plan":[{"type":"configured","id":"MonadRandom-0.5.3-7769c79d6072d028d0cc45baa5b40c1508d3e268","pkg-name":"MonadRandom","pkg-version":"0.5.3","flags":{},"style":"global","pkg-src":{"type":"repo-tar","repo":{"type":"secure-repo","uri":"http://hackage.haskell.org/"}},"pkg-cabal-sha256":"1972a7302aaa409f0b0cd5ff8f5afda03a5156ceb518c2264b2cf0e362123eb6","pkg-src-sha256":"27184dadda0a49abac0208a1e6576b14217a60dc45b6839cd9e90af25ee00a9f","depends":["base-4.14.3.0","mtl-2.2.2","primitive-0.7.4.0-1e3399e34a55b68a47c0fd3ff10d3ddc06c7bb34","random-1.2.1.1-2ac528280489dfc2a07f1e9e00f1f14088d82525","transformers-0.5.6.2","transformers-_-0.7.2-57cc3752e069d87457a495c7fd7d7c2ee03ac3f7"],"exe-depends":[],"component-name":"lib"},{"type":"pre-existing","id":"array-0.5.4.0","pkg-name":"array","pkg-version":"0.5.4.0","depends":["base-4.14.3.0"]},{"type":"pre-existing","id":"base-4.14.3.0","pkg-name":"base","pkg-version":"4.14.3.0","depends":["ghc-prim-0.6.1","integer-gmp-1.0.3.0","rts"]},{"type":"pre-existing","id":"bytestring-0.10.12.0","pkg-name":"bytestring","pkg-version":"0.10.12.0","depends":["base-4.14.3.0","deepseq-1.4.4.0","ghc-prim-0.6.1","integer-gmp-1.0.3.0"]},{"type":"configured","id":"dapp3602022-t4-0.1.0.0-inplace-dapp3602022-t4","pkg-name":"dapp3602022-t4","pkg-version":"0.1.0.0","flags":{},"style":"local","pkg-src":{"type":"local","path":"C:\\Users\\DELL LATITUDE 3550\\Desktop\\Hask\\dapp360_2022_t4\\."},"dist-dir":"C:\\Users\\DELL LATITUDE 3550\\Desktop\\Hask\\dapp360_2022_t4\\dist-newstyle\\build\\x86_64-windows\\ghc-8.10.7\\dapp3602022-t4-0.1.0.0\\x\\dapp3602022-t4","depends":["base-4.14.3.0","random-1.2.1.1-2ac528280489dfc2a07f1e9e00f1f14088d82525","random-shuffle-0.0.4-ea6b2ea4d920844f85c690de568a5207c4228719"],"exe-depends":[],"component-name":"exe:dapp3602022-t4","bin-file":"C:\\Users\\DELL LATITUDE 3550\\Desktop\\Hask\\dapp360_2022_t4\\dist-newstyle\\build\\x86_64-windows\\ghc-8.10.7\\dapp3602022-t4-0.1.0.0\\x\\dapp3602022-t4\\build\\dapp3602022-t4\\dapp3602022-t4.exe"},{"type":"pre-existing","id":"deepseq-1.4.4.0","pkg-name":"deepseq","pkg-version":"1.4.4.0","depends":["array-0.5.4.0","base-4.14.3.0"]},{"type":"pre-existing","id":"ghc-boot-th-8.10.7","pkg-name":"ghc-boot-th","pkg-version":"8.10.7","depends":["base-4.14.3.0"]},{"type":"pre-existing","id":"ghc-prim-0.6.1","pkg-name":"ghc-prim","pkg-version":"0.6.1","depends":["rts"]},{"type":"pre-existing","id":"integer-gmp-1.0.3.0","pkg-name":"integer-gmp","pkg-version":"1.0.3.0","depends":["ghc-prim-0.6.1"]},{"type":"pre-existing","id":"mtl-2.2.2","pkg-name":"mtl","pkg-version":"2.2.2","depends":["base-4.14.3.0","transformers-0.5.6.2"]},{"type":"pre-existing","id":"pretty-1.1.3.6","pkg-name":"pretty","pkg-version":"1.1.3.6","depends":["base-4.14.3.0","deepseq-1.4.4.0","ghc-prim-0.6.1"]},{"type":"configured","id":"primitive-0.7.4.0-1e3399e34a55b68a47c0fd3ff10d3ddc06c7bb34","pkg-name":"primitive","pkg-version":"0.7.4.0","flags":{},"style":"global","pkg-src":{"type":"repo-tar","repo":{"type":"secure-repo","uri":"http://hackage.haskell.org/"}},"pkg-cabal-sha256":"89b88a3e08493b7727fa4089b0692bfbdf7e1e666ef54635f458644eb8358764","pkg-src-sha256":"5b2d6dc2812eb2f6a115f05fcbe3e723d3aeff7894b012c617e075130581add5","depends":["base-4.14.3.0","deepseq-1.4.4.0","template-haskell-2.16.0.0","transformers-0.5.6.2"],"exe-depends":[],"component-name":"lib"},{"type":"configured","id":"random-1.2.1.1-2ac528280489dfc2a07f1e9e00f1f14088d82525","pkg-name":"random","pkg-version":"1.2.1.1","flags":{},"style":"global","pkg-src":{"type":"repo-tar","repo":{"type":"secure-repo","uri":"http://hackage.haskell.org/"}},"pkg-cabal-sha256":"dea1f11e5569332dc6c8efaad1cb301016a5587b6754943a49f9de08ae0e56d9","pkg-src-sha256":"3e1272f7ed6a4d7bd1712b90143ec326fee9b225789222379fea20a9c90c9b76","depends":["base-4.14.3.0","bytestring-0.10.12.0","deepseq-1.4.4.0","mtl-2.2.2","splitmix-0.1.0.4-8f1378aeb544177bd12b9642c7487709620cc097"],"exe-depends":[],"component-name":"lib"},{"type":"configured","id":"random-shuffle-0.0.4-ea6b2ea4d920844f85c690de568a5207c4228719","pkg-name":"random-shuffle","pkg-version":"0.0.4","flags":{},"style":"global","pkg-src":{"type":"repo-tar","repo":{"type":"secure-repo","uri":"http://hackage.haskell.org/"}},"pkg-cabal-sha256":"d0b9783c30208d1ce05caa05139fb00372802edb6c678b73d61b5eb4eebb3aba","pkg-src-sha256":"52704411f040fd0bf2361dad162e35dc13caa6535b2e4908d3513c00a95d0615","components":{"lib":{"depends":["MonadRandom-0.5.3-7769c79d6072d028d0cc45baa5b40c1508d3e268","base-4.14.3.0","random-1.2.1.1-2ac528280489dfc2a07f1e9e00f1f14088d82525"],"exe-depends":[]}}},{"type":"pre-existing","id":"rts","pkg-name":"rts","pkg-version":"1.0.1","depends":[]},{"type":"configured","id":"splitmix-0.1.0.4-8f1378aeb544177bd12b9642c7487709620cc097","pkg-name":"splitmix","pkg-version":"0.1.0.4","flags":{"optimised-mixer":false},"style":"global","pkg-src":{"type":"repo-tar","repo":{"type":"secure-repo","uri":"http://hackage.haskell.org/"}},"pkg-cabal-sha256":"804e2574bc7e32d08cbab91e47ee6287b4df7d50851d73f9e778f94a9a7814c7","pkg-src-sha256":"6d065402394e7a9117093dbb4530a21342c9b1e2ec509516c8a8d0ffed98ecaa","depends":["base-4.14.3.0","deepseq-1.4.4.0"],"exe-depends":[],"component-name":"lib"},{"type":"pre-existing","id":"template-haskell-2.16.0.0","pkg-name":"template-haskell","pkg-version":"2.16.0.0","depends":["base-4.14.3.0","ghc-boot-th-8.10.7","ghc-prim-0.6.1","pretty-1.1.3.6"]},{"type":"pre-existing","id":"transformers-0.5.6.2","pkg-name":"transformers","pkg-version":"0.5.6.2","depends":["base-4.14.3.0"]},{"type":"configured","id":"transformers-_-0.7.2-57cc3752e069d87457a495c7fd7d7c2ee03ac3f7","pkg-name":"transformers-compat","pkg-version":"0.7.2","flags":{"five":false,"five-three":true,"four":false,"generic-deriving":true,"mtl":true,"three":false,"two":false},"style":"global","pkg-src":{"type":"repo-tar","repo":{"type":"secure-repo","uri":"http://hackage.haskell.org/"}},"pkg-cabal-sha256":"044fb9955f63ee138fcebedfdcbe54afe741f2d5892a2d0bdf3a8052bd342643","pkg-src-sha256":"b62c7304c9f3cbc9463d0739aa85cb9489f217ea092b9d625d417514fbcc9d6a","depends":["base-4.14.3.0","ghc-prim-0.6.1","transformers-0.5.6.2"],"exe-depends":[],"component-name":"lib"}]}
